Address 2 Copilot review findings; track 4 deferred deploy fixes
- server/config/reset.js: users.githubid changed from INT to BIGINT
- server/config/auth.js: repeat logins now refresh username/avatarurl/
  accesstoken instead of serving stale data from first signup (role,
  id, created_at untouched)

Also adds a "Before deploying (Phase 9)" section to BUILD_GUIDE.md
tracking the four hardcoded-localhost findings from the same review
(client SERVER_URL, auth.js callbackURL, routes/auth.js redirect,
server.js CORS origin) that are deliberately deferred until there's
an actual deployed environment to configure them against.

+++ server/config/reset.js
@@ -20,7 +20,7 @@ async function createUsersTable() {
   await pool.query(`
     CREATE TABLE users (
       id SERIAL PRIMARY KEY,
-      githubid INT NOT NULL UNIQUE,
+      githubid BIGINT NOT NULL UNIQUE,
       username VARCHAR(100) NOT NULL,
       avatarurl TEXT,
       accesstoken VARCHAR(500),
